% Auto-generated: do not edit by hand
\name{''DashGlobeWrapper}

\alias{''DashGlobeWrapper}

\title{DashGlobeWrapper component}

\description{

}

\usage{
''DashGlobeWrapper(id=NULL, height=NULL, width=NULL)
}

\arguments{
\item{id}{Character. }

\item{height}{Numeric. }

\item{width}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

